Standardize the naming convention of base64-encoded binary data.
I am working on populating our internal (production) authentication
object onto wire in gRPC.  I am going to use Credentials, and because
the authentication object is serialized as binary headers, I will
base64-encode them.

The Credentials object will be passed to gRPC ClientAuthInterceptor and
populated to gRPC Metadata, which accepts binary value (byte[]) when the
key ends with "-bin", and string value otherwise.  Therefore,
ClientAuthInterceptor needs to decode the base64-encoded binary headers
returned by Crendentials, and pass through the others.

ClientAuthInterceptor could use "-bin" as the indicator. However,
because it consumes any Crendentials implementation, it's not safe for
it to rely on this indicator unless it's defined on the Credentials
interface.

This is a documentation-only no-op change.