Spanner use custom thread factory (#5167)
* use custom thread factory for Spanner

Using a custom thread factory for the different underlying gRPC clients
makes it easier to debug potential thread problems, as the name of the
thread indicates to which client it belongs.

* ensure that there are no Spanner threads

* use a managed executor provider for the channel providers

Setting an executor provider on the stub settings would cause the
executors to be shared between the underlying gRPC channels. Instead,
each channel should have its own executor. The executor provider is now
set on the channel provider again, but the executor provider keeps track
of the executors that have been handed out and closes these when the
Spanner instance is closed.

This approach is equal to the approach that was already used for the
spannerWatchdog executor.

* keep track of closed Spanner instances

* added additional test

* increase wait to decrease chance of false negative

* moved check for closed Spanner to separate PR

* base check on actual number of channels

* always use 100 sessions to ensure all channels should be used

* unflake test: do execute more requests to initialize all channels

* giving up test condition, as it is too flaky

The actual number of threads is too dependent on random factors to be
used for testing. Manual testing by calling
ThreadPoolExecutor#prestartAllCoreThreads() shows the desired behavior.

* giving up test assertions as they are too flaky

* ran formatter

* make waiting time dynamic to allow longer waits on slower systems