tests: benchmark should start the transaction (#150)
The steadyIncrease benchmark only checked out a read-only session from
the pool, but did not actually start the transaction. That is only done
once a statement is issued for the transaction.

Note that issuing the statement does not actually execute the
statement. That is deferred to the first call to ResultSet#next().