feat: Allow user-agents to be specified by both internal headers and user headers (#1190)
* feat: Allow user-agents to be specified by both internal headers and user headers

This is primarily meant to address the https://github.com/googleapis/java-bigtable/pull/404#pullrequestreview-480972135. Where both the hbase adapter and the core client needs to set the UserAgent. To reconcile the conflict this PR takes grpc's approach and concatenates the values. The reason this needs to happen now is that for DirectPath we need to ship clients that send a UserAgent, however Bigtable needs to keep stats to differentiate client usage.

* remove stale code