feat: adds support to group logs of the same http request (#257)
* feat: adds support to group logs of the same http request

With TraceLoggingEventEnhancer users can set the traceId of the incoming request and then all the logs that will be created on the same thread will have the traceId property set to that traceId. This allows Cloud Logging to group log entries from the same request in the UI for easier troubleshooting

* feat: adds support to group logs of the same http request

With TraceLoggingEventEnhancer users can set the traceId of the incoming request and then all the logs that will be created on the same thread will have the traceId property set to that traceId. This allows Cloud Logging to group log entries from the same request in the UI for easier troubleshooting

* Fix comment typo

* Update src/main/java/com/google/cloud/logging/logback/TraceLoggingEventEnhancer.java

Co-authored-by: yoshi-code-bot <70984784+yoshi-code-bot@users.noreply.github.com>

Co-authored-by: yoshi-code-bot <70984784+yoshi-code-bot@users.noreply.github.com>