fix: Add cleanup for  Trace context (#266)
* fix: fixing tests to cleanup trace context

Previously we didn't cleanup trace context so it could cause side effects to some tests as a result of added label in trace context tests. To mitigate this, we added a method to cleanup trace context from MDC as well as invoke it before every test

* fix: fixing tests to cleanup trace context

Previously we didn't cleanup trace context so it could cause side effects to some tests as a result of added label in trace context tests. To mitigate this, we added a method to cleanup trace context from MDC as well as invoke it before every test

* fix: lint