chore: simplify BulkWriter (#523)
This is my attempt to simplify some of the callback logic in BulkWriter, and will hopefully allow us add Backoff handling for retries pretty easily.

The overall changes are:

- There are no more batch queues. Instead, there is only ever a single batch. If this batch reaches capacity, it is send.
- To allow 'flush' to block, I am instead keeping a list of pending operations.

One behavior change is that it is now possible that a retry will get scheduled after other operations in a flush. Let's say we have the calls:

- Op1
- Op2
- Flush
- Op3

And Op1 fails. The write order could now be: Op1, Op2, Op3, Op1. This breaks a single unit test, but I think it is sane behavior as it allows us to make progress while we wait for more reponses.

Port of https://github.com/googleapis/nodejs-firestore/pull/1379