feat: Support Array conversion to ResultSet (#326)
* feat: add support for array as ResultSet

* feat: add support for array as ResultSet

JDBC arrays can optionally be converted to ResultSets. This feature was not
implemented for the Cloud Spanner JDBC driver. This would cause DBeaver to
show an error message instead of the actual data when fetching a NUMERIC array.

Other arrays would be fetched correctly by DBeaver, as those array types do not
use the conversion to a ResultSet.

* test: add tests for invalid data types