feat!: Point In Time Recovery (PITR) (#452)
* feat: exposes new pitr-lite database fields

Exposes earliest version time and version retention period fields in the
database class.

* feat: adds it tests for updating version retention

Adds integration tests for updating the version retention period.

* feat: adds new create database tests for pitr

Adds create database tests for PITR and refactors the integration test
class.

* chore: refactors tests

Separates PITR database tests into 2 files for clarity.

* fix: disables pitr-lite tests in emulator

The feature is not supported in the emulator currently.

* fix: closes result set in test

Addresses PR comment.

* fix: updates DatabaseInfo equals/hashcode

To compare version retention period and earliest version time.

* fix: formatting

Fixes formatting of the DatabaseInfo

* feature: adds test for throttled pitr field

Adds test to check for the throttled field in the update database ddl
metadata.

* fix: explain further the pitr-lite params in docs

Adds more explanations to the purpose of the added params for pitr-lite:
version_retention_period and earliest_version_time.

* feat: adds version time to backups

Adds PITR-lite version time to backups. This should make it possible to
specify the consistent time for copying the database.

* test: adds integration tests for pitr backups

* test: adds tests for pitr restore

* test: fixes integration test for pitr restore

* test: fixes backup unit test

* test: fixes npe on pitr backup test

* chore: fixes clirr errors

* chore: refactors / addresses pr comments

* test: fixes the it pitr sad cases tests

* test: fixes pitr backup and restore tests

* test: skips pitr backup and restore tests

This is because the backend for these features is not ready yet.