feat: adds samples for PITR (#837)
* feat: adds samples for pitr

Adds sample for creating a database with a version retention
period. Updates sample for creating a backup with a version time.
Updates sample for restoring a backup, printing the backup version time.

* fix: prints out version time for backup sample

* test: adds test to the pitr create db sample

* test: cleans up after pitr sample test finishes

* test: fixes checkstyle violations

* samples: updates backup samples for pitr

Uses the earliest version time from the database as the version time for
the backup.

* test: fixes pitr samples test

* test: fixes checkstyle violations

* samples: PITR samples backup fix (#921)

* samples: creates a backup using the current time

Instead of using the earliest version time of the database, uses the
current time (from spanner). If we used the earliest version time of the
database instead we would be creating an empty backup, since the
database we are backing up is not 1 hour old (default version retention
period).

* samples: remove unused variable

Clean up create backup samples by removing unused variable.

* samples: parameterises create backup version time (#930)

* samples: fixes spanner samples tests