feat: adds query optimizer statistics support (#385)
* feat: adds optimizer_statistics_package option

Adds the possibility to set the optimizer statistics package when
executing queries. This option can be set in different levels as follows:
1. Through statement hints
2. Through query level configuration
3. Through an environment variable
4. Through application level configuration

If more than one package is set (in different levels) the precedence is from 1 to 4 (1 has top priority).

* test: adds integration tests

Adds integration tests for setting the query options optimizer
statistics package.

* fix: addresses PR comments

Fix missing value in the documentation of the optimizer statistics
package for the connection class.

* fix: adds tests for invalid stats packages

Adds tests for invalid statistics packages (whitespace only ones).

* fix: adds integration tests for query options

Adds an integration test to run the sql script with several expectations
for the query options. These are tests for the optimizer version and
optimizer statistics package.

* fix: formatting of ClientSideStatements.json

This file is using a mix of tabs and spaces. For now I have formatted as
the other lines, so that the diff is clear. In a further PR I will
reformat the whole file to use only spaces.

* tests: fix Connection test

Fixes connection test when using environment variables for retrieving
configurations. This only works when a first connection is created, so
we moved this specific test to its own subclass.

* fix: fix clirr checks

Provides default interface implementations and fixes clirr checks