fix: update CI to use maven 3.8.1 (#662)
Gitbub Action updated LTS(18.04, 20.04) versions of Github-Hosted runners for Ubuntu with the latest version of Apache Maven. The new version of Ubuntu has Apache Maven 3.8.2 which is incompatible with flatten-maven-plugin:1.2.7. This caused all github actions using the latest Ubuntu version to fail with this error.

The immediate solution is to use this community action to install maven 3.8.1 at runtime.