fix: sessions were not always removed from checkedOutSessions (#1438)
Sessions were added to a set of checked out sessions when one was checked out from
the session pool. When the session was released back to the pool, the session would
normally be removed from the set of checked out sessions. The latter would not always
happen if the application that checked out the session did not use the session for
any reads or writes.

Co-authored-by: Thiago Nunes <thiagotnunes@google.com>