chore: support untyped values for statement parameters (#1854)
* feat: support untyped values

* chore: support untyped values

Adds support for untyped values that can be used as statement
parameters. This is required for Spangres, as many native PG drivers
send (some) parameters without an explicit type, and expect the backend
to infer the type from the context. If we were to include a
(placeholder) type for such a parameter, the type inference on the
backend fails.

* chore: support untyped values

Adds support for untyped values that can be used as statement
parameters. This is required for Spangres, as many native PG drivers
send (some) parameters without an explicit type, and expect the backend
to infer the type from the context. If we were to include a
(placeholder) type for such a parameter, the type inference on the
backend fails.

* 🦉 Updates from OwlBot post-processor

See https://github.com/googleapis/repo-automation-bots/blob/main/packages/owl-bot/README.md

Co-authored-by: Owl Bot <gcf-owl-bot[bot]@users.noreply.github.com>