build(fix): update maven config for dependency enforcement of google-http-client-gson (#1631)
In #1615 we started using GsonFactory in some of our tests, as a result dependency enforcement required us to acknowledge the new dependency. Since we were using it in tests, I set the scope to `test`. Setting the scope to `test` actually breaks the transitive dependency resolution performed by the flatten plugin. This results in google-http-client-gson being exclude entirely from the resulting pom and causing downstream consumers to receive a NoClassDefFoundError.

Now, instead of setting the scope to `test` we're explicitly stating that the dependency enforcement plugin should ignore google-http-client-gson.

Related to #1615