chore: cleanup statement execution and test all execute methods of Statement and PreparedStatement (#1321)
* test: add tests for all execute methods of Statement and PreparedStatement

Adds tests for all the different variations of the execute method on
Statement and PreparedStatement. This will allow us to make internal
changes to the way these methods are handled, while keeping control over
any behavior changes that it might introduce.

* chore: share do-with-timeout code (#1324)

Refactor the execute methods a bit so they all share the same
do-with-timeout logic.