fix: re-enable checkstyle in sdk-platform-java-config (#2335)
The checkstyle was initially disabled to address the following error when executing `mvn clean install`:
```
[ERROR] Failed to execute goal org.apache.maven.plugins:maven-checkstyle-plugin:3.3.1:check (checkstyle) on project sdk-platform-java-config: Failed during checkstyle execution: Unable to find configuration file at location: license-checks.xml: Could not find resource 'license-checks.xml'. -> [Help 1]
```

However, since this artifact is meant to be used as a parent for downstream libraries, there is a risk that the `checkstyle.skip` property will be inherited by child poms in the future. This may cause checkstyle to be unintentionally disabled.