feat: improve batching summary errors (#2509)
To avoid data loss, the Batching api throws an exception when the
batcher is closed when at least 1 entry failed. To help debugging, the
BatchingException tries to be helpful by giving some details about the
errors. Since the Batcher lifetime can extend indefinitely, the Batcher
implementation tries to keep a bound on the amount of resources it uses
to track the errors. Previously it would only track exception types and
counts. The idea being that if the customer has the need for fine
grained details, the customer can retrieve the details from the
ApiFuture that was returned when an entry was added. However this hasn't
panned out and creates confusion.

This PR stores a sample of the error messages. The sample is by default
capped to 50 entry and 50 rpc error messages. This can be adjusted by
setting system properties

Thank you for opening a Pull Request! For general contributing
guidelines, please refer to [contributing
guide](https://github.com/googleapis/gapic-generator-java/blob/main/CONTRIBUTING.md)

Before submitting your PR, there are a few things you can do to make
sure it goes smoothly:

- [ ] Make sure to open an issue as a
[bug/issue](https://github.com/googleapis/gapic-generator-java/issues/new/choose)
before writing your code! That way we can discuss the change, evaluate
designs, and agree on the general idea
- [ ] Ensure the tests and linter pass
- [ ] Code coverage does not decrease (if any source code was changed)
- [ ] Appropriate docs were updated (if necessary)

Fixes #<issue_number_goes_here> ☕️