test: support float32 in mock server (#2937)
Adds support for `float32` to the mock server for testing. Also adds a couple of tests that can be used to verify that the mock server supports all types, and that verifies that the client library reads and encodes all types correctly. These tests can be overridden by frameworks using the Java client (e.g. JDBC and PGAdapter) to verify that these also handle these types correctly when used through the API of that tool (I.e. through JDBC or PostgreSQL wire-protocol).