perf: keep comments when searching for params (#2951)
Keep all comments in the SQL string in place when converting positional parameters to named parameters. This reduces the amount of string operations that are needed for each query that is executed, and also enables actually sending comments from the client to Spanner when using positional parameters (e.g. in JDBC).

This is step 3 in the refactoring to share more code between the SpannerStatementParser and PostgreSQLStatementParser.