chore: remove tag restrictions in Connection API (#2970)
The Connection API used a regex to limit the possible values that could
be set for statement and transaction tags. These limitations were not in
sync with what the actual allowed values for tags are in Spanner. The
restrictions in the Connection API have therefore been removed, and all
validation is done by the Spanner backend instead.