fix: update GzipReadableByteChannel to be tolerant of one byte reads (#2512)
If the ReadChannel chunkSize is 0 there will be no library buffering performed during read calls. If gzip decompression support is enabled on the read channel, there is a possibility someone could read fewer than the 4 bytes we initially read.

This change updates the byte tracking to ensure single byte reads are supported.