fix: prepare to generate grafeas (#2761)
In this PR:
- Add suffix to sample source folder
- Add unit tests

Context:
- The package name of grafeas is `io.grafeas.v1` which is not starts
with `com`. When copying samples to destination directory, we need to
change the suffix of source directory.
- grafeas doesn't have a valid api_id (empty string). Before this
change, the `api_id` will be populated as
`{library.api_shortname}.googleapis.com`, which is not correct. `api_id`
should not be populated to `.repo-metadata.json` if set to empty string.

With the library entry added in generation configuration:
```
- api_shortname: containeranalysis
  name_pretty: Grafeas
  product_documentation: https://grafeas.io
  api_description: n/a
  client_documentation: "https://cloud.google.com/java/docs/reference/grafeas/latest/overview"
  release_level: stable
  distribution_name: "io.grafeas:grafeas"
  api_id: ""
  codeowner_team: "@googleapis/aap-dpes"
  library_name: grafeas
  requires_billing: false
  GAPICs:
  - proto_path: grafeas/v1
 ```

The generated java-grafeas is in https://github.com/googleapis/google-cloud-java/pull/10820