fix: Expose Gax meter name (#2865)
This can mitigate the ask to expose `Meter` as a constructor parameter
in `OpenTelemetryMetricsRecorder`. Since `Meter` usually represents the
libary that does the instrumentation, we can not expose it. However, we
can expose the meter name as a public String, so that users of gax can
create Otel views more easily with it.