fix: remove unused DmlBatch span (#3147)
The DmlBatch span that was created for DML batches was never used,
as all traces were registered on the underlying transaction of the
batch. The DmlBatch span was therefore always just an empty
sibling span of the Transaction span, which did contain the traces
of the statements that were executed.