chore: start encapsulating stream and open signaling for BlobDescriptorImpl
Currently, a BlobDescriptorImpl it given its stream. This change shifts things so that BlobDescriptorImpl can create its stream from the requisite inputs. This is needed to allow retries/handling redirects when opening a stream.

This is a first pass in what will be at least a few changes to reach a point where a stream can retry.

Add a test to induce a 404 when attempting to open a BlobDescriptor.