fix: make sure commitAsync always finishes (#3216)
The future that is returned by commitAsync() seems to never finish
in some cases. It is unknown exactly what causes it, and this change
adds a number of safety precautions that should ensure that the
future always returns a result eventually. This is always easier
to debug and handle than a future that never returns a value.