chore: update BlobDescriptor implementation to limit consecutive redirects to 3
* fix: update BlobDescriptor implementation to limit consecutive redirects to 3

If a redirect error is received 4 times in a row, the fourth redirect will raise an error.

Counting is reset when a BidiReadObjectResponse is received.

* chore: de-nestify

* fix: update BlobDescriptorStream to handle BidiReadObjectError

If the BlobDescriptorStream receives a BidiReadObjectError, each read_id contained within will be signaled with the corresponding Status. In the case of a pending future, the future will be resolved with the ApiException corresponding to the Status.

The BlobDescriptorStream will stay "open", and allow future requests be submitted.

In a followup PR I will be implementing finegrained read_id status tracking, allowing for tombstoning an individual read that has not yet been removed for the tracked state. (Removal of tracked state must happen after we have signaled failure to the caller, but the signaling of failure happens on a separate thread to avoid any future callbacks from running on the grpc thread.)

* chore: attach suppressed exception in the event max redirects is exceeded