feat: auto_batch_dml methods + combination with JDBC batching (#1795)
* feat: auto_batch_dml methods + combination with JDBC batching

Adds methods for enabling/disabling auto_batch_dml.

Also enables the use of this feature in combination with other JDBC
batches. This again allows PreparedStatement#addBatch() and
Statement#addBatch(String) to be used in combination with
auto_batch_dml. This again makes it usable in combination with
frameworks like Hibernate.

* test: remove expected error from test

* chore: add clirr diffs