feat: add option for retrying DML as PDML (#3480)
* feat: add option for retrying DML as PDML

Adds an option to the Connection API for automatically retrying DML
statements as Partitioned DML, if the DML statement fails due to
exceeding the Spanner mutation limit. The retry as Partitiond DML
fails if the DML statement is not suitable for Partitioned DML.

The option can be enabled with the `fallback_to_partitioned_dml`
connection variable. This can be set with a SQL statement like this:

```
SET FALLBACK_TO_PARTITIONED_DML = TRUE;
UPDATE my_table SET active=true WHERE true;
```

The property can also be set in the connection URL and by calling
the method `Connection#setFallbackToPartitionedDml(boolean)`.

This option can also be used in the Spanner JDBC driver and
PGAdapter, once those libraries include a version of the Spanner
client that includes this change.

* refactor: include the option in autocommit_dml_mode