chore(deps): Downgrade dev.cel to 0.6.0 (#3441)
`dev.cel` 0.6.0+ brings in `protobuf-java` 4.28.x, which violates the
no-downgrade rule since we still have
[3.25.5](https://github.com/googleapis/sdk-platform-java/blob/e61cbddb3631bc9bef9f07f1fa6be01e219a1457/gapic-generator-java-pom-parent/pom.xml#L34)
in this repo. This is not a major issue now as this dependency is not
used yet in any downstream libraries, but we should downgrade it in case
someone start using it.
Separately, we should prevent this from happening in the future. See
b/382754689 for potential future enhancements.