ci: Add Downstream Protobuf-Java Source Compatibility Test (#3405)
Set up a nightly CI job that tests Protobuf compatibility with
downstream libraries. This step of the CI job is to test source
compatibility.

The nightly job will check the compatibility of these libraries with the
latest in Protobuf-Java 3.x and and 4.x. This job can be manually
invoked with the Protobuf-Java versions as a list.

Future steps may introduce testing for binary compatibility.