fix: always add instance-id for built-in metrics (#3612)
* fix: always add instance-id for built-in metrics

The instance-id was only added for built-in metrics when the headers
of the request was sent. This meant that any request that would NOT
get sent would not include an instance-id, which again would cause
the export of ALL built-in metrics to stop. This would happen because
the metric without an instance-id would remain in the collected metrics.

A request would get into this state if the client ran into a network
issue that would prevent the client from making a network connection
to Spanner.

Fixes #3611

* chore: generate libraries at Sun Jan 26 12:30:46 UTC 2025

---------

Co-authored-by: cloud-java-bot <cloud-java-bot@google.com>