chore: allow generation-input as folder that supplies config and vers… (#3700)
…ions file.

This change adds `--generation-input` option to generation command, only
intend to be used when `--generation-config-path` is not provided. If
none of these 2 provided, then default behavior does not change.

Also added a clean up to delete the temp output folder created in
generation process. This should only be an improvement to local dev
workflow.