fix: plumb mtls endpoint to TransportChannelProvider (#3673)
This PR plumbs the MTLS endpoint (separately from the resolved endpoint)
from the `EndpointContext` to the `InstantiatingGrpcChannelProvider`.

Why not just set the MTLS endpoint in `EndpointContext` if S2A can be
used?
- Although we can decide whether or not to try to use S2A in
`EndpointContext`, we could fail to use S2A in
`InstantiatingGrpcChannelProvider` (if autoconfig doesn't return an
address), in which case we fall back to using a TLS connection
- DirectPath supersedes S2A in `InstantiatingGrpcChannelProvider`, and
the decision to use DirectPath is made in
`InstantiatingGrpcChannelProvider`, not `EndpointContext`. We may decide
to use S2A in `EndpointContext`, but when we go to create the channel in
`InstantiatingGrpcChannelProvider`, we may find that we should be using
DirectPath, in which case we need to use the non-MTLS endpoint