feat: support PostgreSQL isolation level statements (#3706)
* feat: add default_isolation_level connection property

Add a `default_isolation_level` property for the Connection API.
This property will be used by the JDBC driver and PGAdapter to
set a default isolation level for all read/write transactions
that are executed by a connection.

Support for setting an isolation level for a single transaction
will be added in a follow-up pull request.

* feat: specify isolation level per transaction

Add an option to specify the isolation level for a single transaction.
This isolation level overrides the current default isolation level that
has been set for the connection. This option only has an effect for
read/write transactions.

* feat: support PostgreSQL isolation level statements

Adds support for PostgreSQL-dialect SQL statements for setting the isolation
level of a transaction.

* chore: generate libraries at Fri Mar 28 11:51:13 UTC 2025

---------

Co-authored-by: cloud-java-bot <cloud-java-bot@google.com>