feat: Implement X509 certificate provider (#1722)
* feat: Implement X509 certificate provider

This commit introduces the foundational components for workload identity mTLS authentication. It includes classes to load the certificate and private key file paths from a JSON configuration and to provide these credentials as a KeyStore.

This lays the groundwork for future implementation of trust chain handling and subject token generation.

* Address Sonar issues

* Fix lint issues.

* Fix the expected error message in the failing unit test