chore: add jdk 24 to CI tests (#3803)
This is in order to support graalvm upgrade.  

SecurityManager was removed in JDK 24 so had to upgrade to bazel 7.x
which no longer sets command line params to use it. In upgrading to
bazel 7.x also had to replace transitive_deps to use
transitive_compile_time_jars, as that deprecated method was dropped:
https://github.com/bazelbuild/bazel/commit/72058be0628e84741232cb116a7199eb92638f84