fix: getCommitResponse() should return error if tx has not committed (#4021)
getCommitResponse() returned an error if it was called before the transaction
had been committed when using regular sessions. This did not happen when using
multiplexed sessions. This change makes sure that the behavior is the same for
both regular and multiplexed sessions, by adding a check to the
getCommitResponse() method directly.