test: correctly return tx id for update counts (#4045)
The mock Spanner server did not correctly return the transaction ID when
ExecuteStreamingSql was used in combination with an update count result.

Instead of returning the transaction ID of the transaction that was
actually started, it returned the transaction ID that was set on the
request. This would be empty for requests that used a Begin transaction
selector, which again meant that it would return an empty transaction ID.

This change also returns the UPDATE_COUNT field in the result set, which
is also not returned by real Spanner.