chore: Update the template of renovate.json to ignore PR limits (#3902)
We have seen increasingly reports that Renovate Bot is not responsive in
handwritten libraries due to PR rate limits. For example, a dozen PRs
are rate limited in
[java-bigtable](https://github.com/googleapis/java-bigtable/issues/420).

Try to disable the renovate PR limits per
https://docs.renovatebot.com/presets-default/#disableratelimiting. We
tested in
[java-cloud-bom](https://github.com/googleapis/java-cloud-bom/blob/28bec089569e105bfe08d427e8fd7a849b28e472/renovate.json#L2-L3)
repo that it does seems to work. Even though it took renovate bot a day
or so to pick up the configuration.
We'll monitor if it fixes the issue or not and adjust it further if
needed.