feat: add connection property for gRPC interceptor provider (#4149)
* feat: add connection property for gRPC interceptor provider

Add a connection property for setting a gRPC interceptor provider to use for connections.
This allows JDBC and PGAdapter users to set a gRPC interceptor that should be used for
the underlying Spanner client.

This property is a guarded property, as it dynamically invokes the constructor of the
class that is specified in the connection URL. A user must set the Java System property
ENABLE_GRPC_INTERCEPTOR_PROVIDER=true when using this connection property. It should
only be enabled in applications where an untrusted user cannot modify the connection URL
that is being used, as that would allow an untrusted user to dynamically invoke code on
the application host.

* chore: generate libraries at Wed Oct  8 08:52:47 UTC 2025

---------

Co-authored-by: cloud-java-bot <cloud-java-bot@google.com>