chore: Group upper bound dependencies file deps (#3941)
Tested in https://github.com/lqiu96/sdk-platform-java/pull/22

These rules need to be added to the end as renovate `packageRules`
configurations are `mergeable`. This means that it scans the rules
top-down and each subsequent matching rules gets its configurations
overwritten. This tries to make it so that only the `dependencies.txt`
are in its own file.