feat: add unknownLength connection property (#2286)
Adds an `unknownLength` connection property that can be used to configure the length that
the JDBC driver should return as the data type / column length when this is not known.
Spanner does not return the (maximum) length of a column in ResultSetMetadata. This means
that the JDBC driver does not know what the length is of the various columns in query results.

The value of `unknownLength` will be returned when the getPrecision and getColumnDisplaySize
methods of ResultSetMetaData are called.

This connection property aligns with the same connection parameter in the PostgreSQL JDBC driver:
https://jdbc.postgresql.org/documentation/use/#connection-parameters